+++ www/commit-hash.js
@@ -1,1 +1,1 @@
-window.__COMMIT__ = '4b663d34c88c1f63d00c021b2a923b2f1fa444c2';
+window.__COMMIT__ = '31413379be674280b79aabea3eb449c71148c9bf';